Move StrongAlias to its own file
